+++ postcss.config.js
@@ -1,10 +1,5 @@
 module.exports = () => {
   return {
-    plugins: [
-      require('postcss-import'),
-      require('tailwindcss'),
-      require('autoprefixer'),
-      process.env.NODE_ENV === 'production' ? 'postcss-minify' : false,
-    ]
+    plugins: [require('postcss-import'), require('tailwindcss'), require('autoprefixer')],
   };
 };