+++ src/js/defered-module.js
@@ -1,5 +1,5 @@
 const module = () => {
-  console.log('Hello from async loaded module');
+  console.log('Hello from defered module!');
 };
 
 export default module;