+++ src/js/prefetched-module.js
@@ -1,0 +1,5 @@
+const module = () => {
+  console.log('Hello from prefetched module!');
+};
+
+export default module;